+++ src/test/java/runners/Runner.java
@@ -15,7 +15,7 @@ import org.junit.runner.RunWith;
         },
         features = "./src/test/resources/features/ui_features",
         glue = {"stepdefinitions","hooks"},
-        tags = "@us_04",
+        tags = "@password_edition",
         dryRun = false
 )
 public class Runner {
